requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
psycopg==3.2.10
psycopg-binary==3.2.10
python-decouple==3.8
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
wheel==0.45.1
django-model-utils>=4.5.0
